Cleaning of Sonar Issues
As the tags for the code smells of a commit are now the label, the necessary data has been brought in and cleared. Also it has been investigated whether version 1 data can be used for a model by assigning tags based on known issue messages from version 2. It has been shown that this can be done which increases the training data.