Analysis of the relationship between commits and issues, analysis of duplicated issues
When analysing the amount of issues introduced per commit, it could be shown that some commits introduce way more issues than other, which needs to be considered for modelling.
Furthermore, issues are only noted when first detected by SonarQube, with duplications being rare and probably due to reintroduction of errors instead of being systematic.